Add Dry::Validation::PredicateSet and BuiltIn
