+++ Gemfile
@@ -2,6 +2,8 @@ source 'https://rubygems.org'
 
 gemspec
 
+gem 'dry-container', github: 'dryrb/dry-container', branch: 'feature/singleton'
+
 group :tools do
   gem 'faker'
   gem 'rubocop'
