+++ lib/dry/validation.rb
@@ -11,3 +11,4 @@ end
 
 require 'dry/validation/predicate'
 require 'dry/validation/schema'
+require 'dry/validation/predicate_set/built_in'
